web-frontend/api/requirements.txt
requests
numpy
pandas
ta
python-dotenv
openai